Extract mongodb only tests
